Fix "disentanglement" of falling QBlocks
If one or multiple instances of an entangled QBlock group change into a gravity dependent block like sand or gravel and if they are hanging in the air, they will temporarily change into falling blocks. When they re-materialize into solid blocks, they do not re-register in the Entanglement Registry and therefore become "disentangled". Adding this validate(); should fix that behaviour.