Fixed issue #9
This is a very crude solution for this. I had to copy quite some code
because of the huge size of some of the other methods. (Those should
probably be restructured and split up in separate sub-methods.)
Could someone please check if the first two parameters in the method
call in line 622 are ok like this? I haven't quite figured out yet how
the m_currentlyObserved and m_forceObserved variables exactly influence
the behaviour of entangled blocks and groups.