Restructured save system and use Forge logger
QCraft now uses the Forge logging system instead of the Minecraft or
Java one. (in QCraft.java)
Restructured the whole file save system to use the system that is
offered by Forge. This solves several problems with the Entanglements
and Validations of portals sometimes being reset at random moments or on
server restart. (rest of changed files)

This particular version still needs to be play-tested before it can be
released.