Dynamic Quantum Portal sizes 001
-Added config option for maximum portal size (the portal itself, not the
border)
-Added a feature to the Quantum Computer. It will now detect show an
error if it detects multiple possible portal shapes.
Once the QC has determined that there is only one valid portal at some
moment, it will connect to that one and not check for additional portal
possibilities unless it is disconnected from the "connected" portal
somehow. This is not a bug. It reduces lag, if anything.
-Finetuned the "teleport to" coordinates to end up in exactly the middle
of the portal when traveling.