Inter-server item teleports use unique names 004
The dummy item should now work in all cases.
It turns back into the original item if it is transported to a server
that actually does have the original item. (Checked and repaired,
because I derped.)

Something that might be neat to add to this Dummy item, is the ability
for it to transform to the original item when right-clicked if the
original item is now available on this server, however an "inter-server
portal" with an adress pointing to the same server should solve this as
well.